+++ CHANGELOG.md
@@ -6,6 +6,8 @@ follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.2.0-rc13] - 2026-08-07
+
 ### Fixed
 
 - Managed Channel Indexer start no longer builds Basecamp instance IDs that